Maximum Sum Increasing SubSequence
Implementation of Max Sum Increasing Sub Sequence.

Source: https://www.geeksforgeeks.org/dynamic-programming-set-14-maximum-sum-increasing-subsequence/

This problem is a variation of standard Longest Increasing Subsequence (LIS) problem. We need a slight change in the Dynamic Programming solution of LIS problem. All we need to change is to use sum as a criteria instead of length of increasing subsequence.